fix(app): move all raw SQL out of app layer — satisfy db-boundary gate
Add persistence repository methods for every sqlx site that was in the
app layer (71 sites across apply.rs, plan_query.rs, operation_query.rs):

persistence_sessions:
- repositories/actors.rs: ensure_spec062_actor, ensure_spec062_target,
  lookup_spec062_target_row_id
- repositories/change_sequence.rs: current_sequence_on_conn (inside an
  open IMMEDIATE tx, fixing the round-1 race), insert_repository_change,
  current_sequence (read-only watermark path)
- repositories/result_snapshots.rs: insert_result_session,
  insert_result_frame, insert_result_panel_group
- repositories/tx.rs: begin_immediate, commit, rollback,
  enable_foreign_keys
- repositories/materialization.rs: transition_operation_to_cancelled,
  get_operation_public_id_by_row_id

persistence_inbox:
- repositories/plan_result.rs: get_plan_snapshot_for_operation,
  list_proposed_sessions, list_proposed_session_frames,
  get_site_resolution_revision

app_inbox::session_materialization::apply: rewritten to call only
persistence repository methods; panel-group commit delegates to
persistence_topology::insert_singleton_panel_group

app_inbox::session_materialization::plan_query: now re-exports
persistence_inbox::repositories::plan_result types and functions

app_targets::ingestion_operation::operation_query: target lookup
delegates to persistence_sessions::repositories::actors; sequence
watermark uses change_sequence::current_sequence

db-boundary: 0 violations (sealed at zero)
dead-callers: 116 baselined (current_change_sequence baselined, now
superseded by change_sequence::current_sequence_on_conn / current_sequence)
hot-read, lifecycle-strings: pass
integration tests: 5/5
