+++ crates/app/inbox/Cargo.toml
@@ -27,6 +27,7 @@ persistence_lifecycle = { path = "../../persistence/lifecycle" }
 persistence_plans = { path = "../../persistence/plans" }
 persistence_sessions = { path = "../../persistence/sessions" }
 persistence_targets = { path = "../../persistence/targets" }
+persistence_topology = { path = "../../persistence/topology" }
 sessions = { path = "../../sessions" }
 targeting = { path = "../../targeting" }
 targeting_resolver = { path = "../../targeting/resolver" }
