+++ crates/persistence/inbox/src/repositories/mod.rs
@@ -2,4 +2,5 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 pub mod inbox;
+pub mod plan_result;
 pub mod q_inbox;
