+++ crates/persistence/sessions/src/repositories/mod.rs
@@ -1,9 +1,13 @@
 // Copyright (C) 2024-2026 Sjors Robroek
 // SPDX-License-Identifier: AGPL-3.0-only
 
+pub mod actors;
 pub mod calibration_sessions;
+pub mod change_sequence;
 pub mod equipment_resolution;
 pub mod materialization;
 pub mod metadata_resolution;
+pub mod result_snapshots;
 pub mod sessions;
 pub mod supersession;
+pub mod tx;
